Replaced Delete, Save, and Update API integration tests with 1 test: GranulesEndpointSpec
